deeplearning/requirements.txt
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-Script==2.0.6
Flask==1.0.3
Jinja2==2.10.1
SQLAlchemy==1.3.4
Werkzeug==0.15.5
alembic==1.0.10
flasgger==0.9.2
itsdangerous==1.1.0
pytz==2019.1
requests==2.22.0
setproctitle==1.1.10
markupsafe==2.0.1
WTForms==3.0.1
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
Pillow==10.0.1